Extend filter "objectId" to "entityId" and "repositoryId"
Before "objectId" only filtered events which are directly connected to
the uuid. However for example a CheckoutRevision is not included into
the filter for an entity since it's main target uuid is the revision and
not the entity. This quickfix extended the filter to "entityId" and
"repositoryId" as well. However this will only be a short fix since we
want to pass the filter to the database layer anyway.