Remove double quotes from values in env variables
This will enable the usage of the `.env` file with `docker run` and the `--env-file` option, since the docker run subcommand considers surrounding quotes as part of the value of environment variables.

Since variable expansion is not needed for the values of the environment variables, the double quotes can be removed.

Tested that the file can still be sourced with `source .env`.

Fixes https://github.com/serlo/api.serlo.org/issues/909