refactor!(user): remove mutation setEmail
It just changes the email in mysql database, but not in the kratos db,
what would lead to error when trying to login with email or asking
for changing password.
Changing the email is also something that happens quite rare, so we
don't need to support this feature at the moment, since any change can
be done manually.